requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
gevent>=21.8.0
werkzeug>=2.0.0
streamlit>=1.31.0
plotly>=5.18.0
voyageai>=0.1.0
pinecone-client>=3.0.0
datasets>=2.16.0
sentence-transformers>=2.3.0
gunicorn>=20.1.0
